:recycle:[#1823] move steps to individual files
