+++ src/openzaak/config/setup_configuration/steps/__init__.py
@@ -1,0 +1,2 @@
+from .demo import DemoUserStep
+from .selectielijst import SelectielijstAPIConfigurationStep
